Fixed Edit Bug And Badly Started AnswerSpawner
Edit button bug was easy to fix. Needed to break the loop in the
MainMenu method which was causing the edit buttons ahead of the one
clicked to enter after being told to leave. Breaking the loop prevented
the loop from telling the edit buttons to reenter.

I'll fix up AnswerSpawner when I don't have a huge headache and a lack
of will...